pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "emp-agents"
version = "0.1.5"
description = "A Python library for building low-code, capable and extensible autonomous agent systems."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "Johnny"},
  {name = "Rick"}
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent"
]
keywords = ["emp", "agents", "python"]

dependencies = [
    "aiohttp==3.10.5",
    "anthropic==0.34.0",
    "cryptography==43.0.3",
    "eth-rpc-py",
    "eth-typeshed-py",
    "fast-depends>=2.4.0",
    "httpx==0.27.2",
    "pydantic==2.9.2",
    "tiktoken==0.7.0",
    "fast_depends>=2.4.0",
]

[project.optional-dependencies]
tools = [
    "tweepy>=4.14.0"
]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0"
]
dev = [
    "jupyter==1.1.1"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["emp_agents*"]

[tool.setuptools.package-dir]
emp_agents = "src/emp_agents"

[tool.black]
line-length = 88
target-version = ['py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_third_party = ["pytest", "black", "isort"]
default_section = "THIRDPARTY"
